rename engines/toastEngines.js to logic/toasts.js - remove duplicate stats context menu
